Move validation logic to utils.js
